+++ routers/produit.js
@@ -17,8 +17,8 @@ router.post('/add',upload,async (req,res)=>{
          });
         let results= produit_validation.validate(req.body);
         if(results.error)
-            return res.status(403).send(results.error.details[0].message);
-        res.status(200).send(await product.save());
+            return res.status(200).json({ message : results.error.details[0].message });
+        res.status(200).json(await product.save());
     } catch (error) {
         res.status(500).send('Error saving Product :'+error.message);
     }
@@ -28,7 +28,7 @@ router.post('/add',upload,async (req,res)=>{
 router.get('/:id',async (req,res)=>{
     let produit = await Produit.findById(req.params.id)
     if (!produit){
-          return res.status(404).send("Product Not Exist");
+          return res.status(200).json({ message : "Product Not Exist" });
         }else{
             res.status(200).send(produit);    
         }
@@ -39,7 +39,7 @@ router.get('/:id',async (req,res)=>{
 router.get('',async (req,res)=>{
     try {
         let produits = await Produit.find();
-        res.status(200).send(produits)
+        res.status(200).json(produits)
     } catch (error) {
         res.status(500).send('Error get All Products :'+error.message);
     }
@@ -51,22 +51,26 @@ router.put('/:id',async (req,res)=>{
     try {
         let results= produit_validation.validate(req.body);
         if(results.error)
-            return res.status(403).send(results.error.details[0].message);
+            return res.status(200).json({ message : results.error.details[0].message });
         
         await Produit.updateOne({_id : req.params.id}, req.body);
-        res.send(await Produit.findById(req.params.id));
+        res.status(200).json(await Produit.findById(req.params.id));
     } catch (error) {
         res.status(500).send('Error updating Product :'+error.message);
     }
     
 });
 // Delete Product
-router.delete('/:id',async (req,res)=>{
+router.delete('delete/:id',async (req,res)=>{
+    var ObjectId = require('mongoose').Types.ObjectId;
+    if(!ObjectId.isValid(req.params.id)){
+        return res.status(200).json({ message : "Product Not Exist" });
+    }
     try {
         let produit = await Produit.findByIdAndRemove(req.params.id);
         if(!produit)
-            return res.status(404).send('Product with id is not found');
-        res.send(produit);
+            return res.status(200).json({ message : 'Product with id is not found' });
+        res.status(200).json(produit);
     } catch (error) {
         res.status(500).send('Error Deleting Product :'+error.message);
     }
